Add setup interface to install required fonts
This commit adds a setup interface to the metanorma CLI. This
interface should be run to setup all required dependencies. For
now we plugging in the fonts installation process.

This is still very early, and very simplified but the interface
will stay the same and we will make the core functionality more
modular in the coming days.